Added new BaseWebViewActivity
