+++ src/main/java/com/deuna/maven/webviews/ElementsActivity.kt
@@ -1,11 +1,8 @@
 package com.deuna.maven.webviews
 
-import ElementsResponse
 import android.os.*
-import com.deuna.maven.*
 import com.deuna.maven.element.domain.*
 import com.deuna.maven.shared.*
-import org.json.*
 
 class ElementsActivity() : BaseWebViewActivity() {
 
@@ -13,7 +10,7 @@ class ElementsActivity() : BaseWebViewActivity() {
         const val EXTRA_URL = "EXTRA_URL"
         private var callbacks: ElementsCallbacks? = null
 
-        public fun setCallbacks(callbacks: ElementsCallbacks) {
+        fun setCallbacks(callbacks: ElementsCallbacks) {
             this.callbacks = callbacks
         }
     }
@@ -32,9 +29,12 @@ class ElementsActivity() : BaseWebViewActivity() {
 
     }
 
-
     override fun getBridge(): WebViewBridge {
-        return ElementsBridge()
+        return ElementsBridge(
+            context = this,
+            callbacks = callbacks,
+            closeEvents = closeEvents
+        )
     }
 
     override fun onNoInternet() {
@@ -54,52 +54,4 @@ class ElementsActivity() : BaseWebViewActivity() {
         super.onDestroy()
     }
 
-    inner class ElementsBridge() : WebViewBridge() {
-        override fun handleEvent(message: String) {
-            try {
-                val json = JSONObject(message)
-                val eventData = ElementsResponse.fromJson(json)
-
-                callbacks?.eventListener?.invoke(eventData.type, eventData)
-                when (eventData.type) {
-
-                    ElementsEvent.vaultSaveSuccess, ElementsEvent.cardSuccessfullyCreated -> {
-                        callbacks?.onSuccess?.invoke((eventData))
-                    }
-
-                    ElementsEvent.vaultFailed, ElementsEvent.cardCreationError, ElementsEvent.vaultSaveError -> eventData.data.metadata?.let {
-                        handleError(
-                            eventData
-                        )
-                    }
-
-                    ElementsEvent.vaultClosed -> {
-                        closeElements(this@ElementsActivity)
-                        callbacks?.onCanceled?.invoke()
-                    }
-
-                    else -> {
-                        DeunaLogs.debug("ElementsBridge Unhandled event: ${eventData.type}")
-                    }
-                }
-                eventData.let {
-                    if (closeEvents.contains(it.type)) {
-                        closeElements(this@ElementsActivity)
-                    }
-                }
-            } catch (e: Exception) {
-                DeunaLogs.debug("ElementsBridge JSONException: $e")
-            }
-        }
-
-        private fun handleError(response: ElementsResponse) {
-            callbacks?.onError?.invoke(
-                ElementsError(
-                    ElementsErrorType.VAULT_SAVE_ERROR,
-                    response.data.user
-                )
-            )
-        }
-
-    }
 }
